arch.sh: setting hostname defaults to existing

+++ scripts/arch.sh
@@ -83,9 +83,11 @@ locale-gen
 
 echo 'LANG=en_US.UTF-8' > /etc/locale.conf
 
-read -p 'hostname: ' hostname
-echo $hostname > /etc/hostname
-hostname $hostname  # keygen.sh won't read /etc/hostname until reboot
+read -p "hostname: [$(hostname)] " hostname
+if [ -n "$hostname" ]; then
+  echo $hostname > /etc/hostname
+  hostname $hostname  # keygen.sh won't read /etc/hostname until reboot
+fi
 
 echo 'Root password'
 passwd root
